fix : 예약내역 목록 조회 state list로 받아올 수 있게 수정

+++ src/main/java/com/airbng/dto/reservation/ReservationSearchResponse.java
@@ -22,7 +22,7 @@ public class ReservationSearchResponse {
     private String startTime;
     private String endTime;
     private String dateOnly;
-    private Integer durationHours;
+    private Double durationHours;
 //    private String itemNames; // 예: "가방, 신발"
     private List<JimTypeResult> jimTypeResults;
 }
